Combine WPEViewClient and WebChromeClient
There is no reason to have two client interfaces for WPEView.
Move WebChromeClient interface to WPEViewClient to avoid confusion
with two client interfaces. Android System WebView also has client
interface called WebChromeClient so this renaming also avoids confusing
WPEView client with system webview client interface.
